% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_observation_window.R
\name{add_observation_window}
\alias{add_observation_window}
\title{Function to add observations from one data frame to another}
\usage{
add_observation_window(df, df2, threshold = 10, n_max = 60)
}
\arguments{
\item{df}{Data frame to add observations to}

\item{df2}{Data frame to extract observation window}

\item{threshold}{Minimum gap between observations in df (seconds). Used for flagging. Observations closer than threshold with be flagged as 1}

\item{n_max}{Maximum number of seconds in window before and after each observation in df. Used to prevent windows becoming too large if observations are far apart}
}
\value{
Tibble containing nested observations from df2
}
\description{
Function to add observations from one data frame to another according to a time window
}
\author{
Shona Wilde
}
