% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_optimum_offset.R
\name{apply_optiumum_offset}
\alias{apply_optiumum_offset}
\title{Function to apply optimum offset to align peaks}
\usage{
apply_optiumum_offset(
  df,
  cols,
  align_by = "nox",
  offset_variable = "value",
  n = 10
)
}
\arguments{
\item{df}{Data frame to containing observations to align}

\item{cols}{Columns to align, specified as in \code{\link[tidyr]{tidyr_tidy_select}}}

\item{align_by}{Column used to align all other variables to}

\item{offset_variable}{Name of variable to align}

\item{n}{Range of offsets to apply. Varies from -n to +n}
}
\value{
Tibble containing aligned observations
}
\description{
Function to apply optimum offset to align peaks of different variables
}
\author{
Shona Wilde
}
