% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rolling_background.R
\name{calculate_rolling_background}
\alias{calculate_rolling_background}
\title{Function to calculate background values based on a rolling percentile.}
\usage{
calculate_rolling_background(
  df,
  cols,
  width,
  probs = 0.01,
  align = "center",
  na.rm = TRUE,
  round = NA
)
}
\arguments{
\item{df}{Input data}

\item{cols}{Columns to calculate background for}

\item{width}{Width of rolling window}

\item{probs}{Percentile to calculate. Defualt is 1st}

\item{align}{Should the index of the result be center (default), left or right aligned compared to the rolling window of observations?}

\item{na.rm}{Should NAs be removed before the percentile is computed?}

\item{round}{Number of digits to round result to}
}
\value{
Data frame with additional background variables
}
\description{
Function to calculate background values based on a rolling percentile.
}
\examples{
# calculate rolling 5-day 1st percentiles of ozone and solar radiation
calculate_rolling_background(
 df = airquality,
 cols = c(Ozone, Solar.R),
 width = 5,
 probs = 0.01
)

}
