% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raster_df.R
\name{clean_raster_df}
\alias{clean_raster_df}
\title{Tidy output of raster summary}
\usage{
clean_raster_df(df)
}
\arguments{
\item{df}{Data frame resulting from summarising a raster object using \code{\link{terra::global}}}
}
\value{
Tibble
}
\description{
This function is specific to the era5 reanalysis dataset
}
\author{
Shona Wilde
}
