% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_weighted_quant_reg.R
\name{distance_weighted_quant_reg}
\alias{distance_weighted_quant_reg}
\title{Function to perform distance-weighted quantile regression}
\usage{
distance_weighted_quant_reg(
  data,
  x = "x",
  formula,
  distance,
  tau = c(0.5, 0.75, 0.9, 0.95, 0.99),
  sigma = 20
)
}
\arguments{
\item{data}{Input data.}

\item{x}{Variable for calculating weights.}

\item{formula}{A formula specifying the model, as in \code{\link[quantreg]{rq}} and \code{\link[stats]{lm}}.}

\item{distance}{Distance variable.}

\item{tau}{The quantile(s) to be estimated. Generally a number between 0 and 1.}

\item{sigma}{Width of the Gaussian kernel smoothing function.
Small \sigma leads to a narrow Gaussian and hence allows the user to focus on very localised effects, whereas
a bigger \sigma has the effect of smoothing things out.}
}
\value{
Tibble containing model results.
}
\description{
Function to perform distance-weighted quantile regression
}
\author{
Shona Wilde
}
