% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_run_regression.R
\name{do_run_regression}
\alias{do_run_regression}
\title{Function to perform run regression analysis}
\usage{
do_run_regression(
  df,
  formula,
  n = 5,
  complete = T,
  method = "SMA",
  verbose = T
)
}
\arguments{
\item{df}{Input data frame.}

\item{formula}{A formula specifying the model, as in \code{\link[lmodel2]{lmodel2}} and \code{\link[stats]{lm}}.}

\item{n}{Size of moving window.}

\item{complete}{Should the function be evaluated on complete windows only?
If FALSE then partial computations will be allowed.}

\item{method}{Regression method as in \code{\link[lmodel2]{lmodel2}}.
Options are ordinary least squares (OLS), major axis (MA), standard major axis (SMA), and ranged major axis (RMA).}

\item{verbose}{Should messages be displayed to the user?}
}
\value{
Tibble of original data with list columns containing observations for each moving window, model objects and model summaries.
}
\description{
\code{do_run_regression} calculates the regression results for fixed-width, overlapping windows of data.
}
\examples{
\dontrun{

# Perform rolling regression with window size 3
do_run_regression(df, nox~co2, n = 3)

}


}
\author{
Shona Wilde
}
