% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_run_regression2.R
\name{do_run_regression2}
\alias{do_run_regression2}
\title{Function to apply a rolling linear regression model to data}
\usage{
do_run_regression2(df, formula, n = 5, complete = T, verbose = T)
}
\arguments{
\item{df}{Input data frame.}

\item{formula}{A formula specifying the model, as in code{\link[stats]{lm}}.}

\item{n}{Length of moving window.}

\item{complete}{Should the function be evaluated on complete windows only?
If FALSE then partial computations will be allowed.}

\item{verbose}{Should messages be displayed to the user?}
}
\value{
Tibble with list columns containing observations and model summaries for each window.
}
\description{
\code{do_run_regression} calculates the regression results for fixed-width, overlapping windows of data.
}
\examples{
\dontrun{

# Perform rolling regression with window size 3
do_run_regression(df, nox~co2, n = 3)

}


}
\author{
Shona Wilde
}
