% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model_coeff.R
\name{extract_model_coeff}
\alias{extract_model_coeff}
\title{Function to extract coefficients from a GAM model}
\usage{
extract_model_coeff(model)
}
\arguments{
\item{model}{Model object.}
}
\value{
Tibble containing model results.
}
\description{
Function to extract coefficients from a GAM model
}
\author{
Shona Wilde
}
