% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_raster_by_date.R
\name{filter_raster_by_date}
\alias{filter_raster_by_date}
\title{Filter layers of a raster brick between two dates}
\usage{
filter_raster_by_date(raster, start, end)
}
\arguments{
\item{raster}{Raster brick for filtering. Layers of the brick must correspond to dates e.g hourly time series data}

\item{start}{Start date (yyyy-mm-dd)}

\item{end}{End date (yyyy-mm-dd)}
}
\value{
Filtered raster object
}
\description{
Function to subset layers of a raster brick object based on a date range. \code{filter_raster_by_date} was designed for use on the era5 reanalysis dataset.
}
\author{
Shona Wilde
}
