% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rolling_observations.R
\name{get_rolling_observations}
\alias{get_rolling_observations}
\title{Function to get observations resulting from a rolling window}
\usage{
get_rolling_observations(df, n = 5)
}
\arguments{
\item{df}{Input data frame.}

\item{n}{Size of rolling window.}
}
\value{
Tibble.
}
\description{
Function to get observations resulting from a rolling window
}
\author{
Shona Wilde & Stuart Grange
}
