% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haversine_distance.R
\name{haversine_distance}
\alias{haversine_distance}
\title{Haversine distance}
\usage{
haversine_distance(long_1, lat_1, long_2, lat_2)
}
\arguments{
\item{long_1}{Start longitude}

\item{lat_1}{Start latitude}

\item{long_2}{End longitude}

\item{lat_2}{End latitude}
}
\value{
Numeric value
}
\description{
Haversine distance
}
\author{
Shona Wilde & David Carslaw
}
