% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gpx.R
\name{plot_gpx}
\alias{plot_gpx}
\title{Function to plot GPX files on a map}
\usage{
plot_gpx(file)
}
\arguments{
\item{file}{GPX file}
}
\description{
Function to plot GPX files on a map
}
\author{
Shona Wilde
}
