% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sf_leaflet.R
\name{plot_sf_leaflet}
\alias{plot_sf_leaflet}
\title{Function to plot spatial objects with an interactive \strong{leaflet} plot.}
\usage{
plot_sf_leaflet(sf, colour = "blue", popup = TRUE, transform = TRUE)
}
\arguments{
\item{sf}{A \strong{sf} spatial object.}

\item{colour}{Colour of geometries.}

\item{popup}{Should interactive popups be included in the map if possible?}

\item{transform}{Should the projection system be transformed to WGS84 before
plotting?}
}
\value{
Leaflet map/plot.
}
\description{
Function to plot spatial objects with an interactive \strong{leaflet} plot.
}
\author{
Stuart K. Grange & Shona Wilde
}
