% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpx.R
\name{read_gpx}
\alias{read_gpx}
\title{Read GPX}
\usage{
read_gpx(file, lines = T)
}
\arguments{
\item{file}{File to read. Must have a .gpx extension}

\item{lines}{Convert point geometries to lines?}
}
\value{
\code{sf} data frame
}
\description{
Read a GPX file into a spatial data frame
}
\author{
Shona Wilde
}
