% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sample_repeats.R
\name{slice_sample_repeats}
\alias{slice_sample_repeats}
\title{Sample random rows}
\usage{
slice_sample_repeats(df, n, times, replace = F)
}
\arguments{
\item{df}{A data frame or tibble}

\item{n}{Sample size}

\item{times}{Number of repeated samples to take}

\item{replace}{Should sampling be performed with (\code{TRUE}) or without (\code{FALSE}) replacement?}
}
\value{
Tibble
}
\description{
Create multiple sub-samples of a chosen size containing rows of a data frame or tibble
}
\author{
Shona Wilde
}
