% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_arrange_points.R
\name{st_arrange_points}
\alias{st_arrange_points}
\title{Function to arrange a set of points}
\usage{
st_arrange_points(st, direction = "clockwise")
}
\arguments{
\item{st}{Object of class \code{sf}.}

\item{direction}{Direction to order points in. Options are 'clockwise', 'anticlockwise', 'x', or 'y'.}
}
\value{
\code{st} with ordered geometry.
}
\description{
\code{st_arrange_points} can be used to arrange a set of points/coordinates in either a clockwise or anticlockwise direction or in ascending order according
to their latitude or longitude.
}
\author{
Shona Wilde
}
