% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_average_linestring.R
\name{st_average_linestring}
\alias{st_average_linestring}
\title{Function to average groups of linestrings}
\usage{
st_average_linestring(st, distance = 5, crs = 4326)
}
\arguments{
\item{st}{\code{sf} object containing LINESTRING geometries.}

\item{distance}{Resolution to perform interpolation at, in m.}

\item{crs}{Projection of output. Default is to convert to WGS 84.}
}
\value{
\code{sf} object containing summarised geometries.
}
\description{
Calculate the average path of linestrings. Designed for use with the road network where a single named road may be
made up of multiple linestrings.
}
\author{
Shona Wilde
}
