% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bbox_from_df.R
\name{st_bbox_from_df}
\alias{st_bbox_from_df}
\title{Function for promoting a data frame containing bbox coordinates to a spatial data frame}
\usage{
st_bbox_from_df(df, crs = 4326)
}
\arguments{
\item{df}{Data frame containing bbox boundaries in columns. Order of columns must be xmin, ymin, xmax, ymax}

\item{crs}{Coordinate reference system of df}
}
\value{
\code{sf}object containing POLYGON of bounding box
}
\description{
Function for promoting a data frame containing bbox coordinates to a spatial data frame
}
\author{
Shona Wilde
}
