% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_calculate_background.R
\name{st_calculate_background}
\alias{st_calculate_background}
\title{Calculate background values from spatial data}
\usage{
st_calculate_background(st, st_point, formula, sigma, verbose = T)
}
\arguments{
\item{st}{sf object containing spatial data}

\item{st_point}{Locations to perform calculation at. Should be an sf object containing POINT geometries}

\item{formula}{Formula used for regression. Background values are calculated for each variable in the formula}

\item{sigma}{Width of Gaussian kernel in m}

\item{verbose}{Display message to user?}
}
\value{
sf object
}
\description{
This function is designed to be used prior to \code{st_weighted_quantile_regression}.
Background values are calculated using Gaussian kernel weighting from a point of interest. Observations are weighted according to their distance away from the point of interest.
The median of the lowest 20 observations at each point is taken as the background value.
}
\author{
Shona Wilde
}
