% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_calculate_lagged_distance.R
\name{st_calculate_lagged_distance}
\alias{st_calculate_lagged_distance}
\title{Function to calculate distance between consecutive points}
\usage{
st_calculate_lagged_distance(
  st,
  name = "distance_from_previous",
  as_numeric = T,
  crs = NA
)
}
\arguments{
\item{st}{Object of class \code{sf}.}

\item{name}{Name of new column.}

\item{as_numeric}{Should the new column be numeric? Default is TRUE.}

\item{crs}{Projection of output. Default is to leave the projection of code{st} unchanged.}
}
\value{
An object of the same class as \code{st}.
}
\description{
Function to calculate distance between consecutive points
}
\author{
Shona Wilde
}
