% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_calculate_xy_distances.R
\name{st_calculate_xy_distances}
\alias{st_calculate_xy_distances}
\alias{st_calculate_xy_distance}
\title{Calculate horizontal and vertical distances between points}
\usage{
st_calculate_xy_distances(st, id = NA)

st_calculate_xy_distance(location1, location2)
}
\arguments{
\item{st}{Spatial data frame containing points}

\item{id}{ID variable for each point e.g name. If no ID is supplied row numbers are used}

\item{location1}{Start point (sf object containing single point)}

\item{location2}{End point (sf object containing single point)}
}
\value{
Tibble
}
\description{
Calculate horizontal and vertical distances between points
}
\author{
Shona Wilde
}
