% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_combine_lines.R
\name{st_combine_lines}
\alias{st_combine_lines}
\title{Function to combine multiple linestrings,}
\usage{
st_combine_lines(st)
}
\arguments{
\item{st}{\code{sf} LINES object.}
}
\value{
\code{sf} object containing a single LINESTRING.
}
\description{
Function to combine multiple linestrings,
}
\author{
Shona Wilde
}
