% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_coords.R
\name{st_coords}
\alias{st_coords}
\title{Retrieve coordinates in tibble form}
\usage{
st_coords(st, names = c("x", "y"))
}
\arguments{
\item{st}{Object of class sf}

\item{names}{Vector of names for the columns in the order x, y}
}
\value{
Tibble of x and y coordinates
}
\description{
Retrieve coordinates in tibble form
}
\author{
Shona Wilde
}
