% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_coords_as_point.R
\name{st_coords_as_point}
\alias{st_coords_as_point}
\title{Covert a coordinate pair to a point}
\usage{
st_coords_as_point(latitude, longitude, crs = 4326)
}
\arguments{
\item{latitude}{Latitude}

\item{longitude}{Longitude}

\item{crs}{Coordinate reference system of supplied latitude and longitude}
}
\value{
\code(sf) object with POINT geometry
}
\description{
Covert a coordinate pair to a point
}
\author{
Shona Wilde
}
