% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_from_df.R
\name{st_from_df}
\alias{st_from_df}
\title{Function for promoting a data frame to an \code{sf} object}
\usage{
st_from_df(
  df,
  latitude = "lat",
  longitude = "long",
  remove_empty = F,
  crs = 4326
)
}
\arguments{
\item{df}{Data frame to be converted to spatial data frame - currently only works for \code{POINTS}}

\item{latitude}{Latitude variable in df}

\item{longitude}{Longitude variable in df}

\item{remove_empty}{Should rows with missing geometries be removed? Default is FALSE.}

\item{crs}{Coordinate reference system of df}
}
\value{
\code{sf} object
}
\description{
Function for promoting a data frame to an \code{sf} object
}
\author{
Shona Wilde
}
