% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_get_boundary.R
\name{st_get_boundary}
\alias{st_get_boundary}
\title{Function to return bounding box of spatial geometries in a spatial data frame}
\usage{
st_get_boundary(st, group_by, buffer = 0, crs = 4326)
}
\arguments{
\item{st}{Object of class \code{sf}.}

\item{group_by}{Optional grouping variable.}

\item{buffer}{Buffer distance in m.}

\item{crs}{Projection of output. Default is to convert to \emph{WGS 84}.}
}
\value{
An object of the same class as \code{st}, with manipulated geometry.
}
\description{
Function to return bounding box of spatial geometries in a spatial data frame
}
\author{
Shona Wilde
}
