% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_interpolate.R
\name{st_interpolate}
\alias{st_interpolate}
\title{Function to interpolate points along a line geometry}
\usage{
st_interpolate(st, distance, crs = 4326)
}
\arguments{
\item{st}{Object with class \code{sf} or \code{sfc}. The geometry type must be a 'LINESTRING'.}

\item{distance}{Distance between points in m.}

\item{crs}{Projection of output. Default is to convert to \emph{WGS 84}.}
}
\value{
An object of the same class as \code{st}, with manipulated geometry.
}
\description{
Return points at equal distances along a line.
}
\author{
Shona Wilde
}
