% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_make_tidygraph.R
\name{st_make_tidygraph}
\alias{st_make_tidygraph}
\title{Function to create a \code{tbl_graph} from spatial linestrings}
\usage{
st_make_tidygraph(st, directed = F)
}
\arguments{
\item{st}{\code{sf} object with \code{LINESTRING} geometries.}

\item{directed}{Should the graph be directed? Default is FALSE.}
}
\value{
\code{tbl_graph}.
}
\description{
This function is adapted from \href{https://r-spatial.org/r/2019/09/26/spatial-networks.html}{Spatial networks in R with sf and tidygraph}.
}
\seealso{
\code{\link[tidygraph]{tbl_graph}}
}
\author{
Shona Wilde
}
