% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_points_to_lines.R
\name{st_points_to_lines}
\alias{st_points_to_lines}
\title{St points to lines}
\usage{
st_points_to_lines(st, group_by = NA, do_union = F)
}
\arguments{
\item{st}{Object with class \code{sf} or \code{sfc}. The geometry type must be 'POINTS'}

\item{group_by}{Optional grouping variable}

\item{do_union}{Should geometries be created by unioning using st_union, or simply by combining using st_combine?}
}
\value{
Sf object with LINESTRING geometries
}
\description{
Wrapper to cast a spatial points object to lines
}
\author{
Shona Wilde
}
