% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_projection.R
\name{st_projection}
\alias{st_projection}
\title{Function to return the projection system of a spatial object.}
\usage{
st_projection(st)
}
\arguments{
\item{st}{Object with class \code{sf} or \code{sfc}.}
}
\value{
Character.
}
\description{
Function to return the projection system of a spatial object.
}
\author{
Shona Wilde
}
