% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_snap_points_to_lines.R
\name{st_snap_points_to_lines}
\alias{st_snap_points_to_lines}
\title{Function to "snap" spatial points to the nearest spatial line.}
\usage{
st_snap_points_to_lines(x, y, max_dist = 1000)
}
\arguments{
\item{x}{\code{sf} POINTS object}

\item{y}{\code{sf} LINES object to snap points to.}

\item{max_dist}{The maximum distance allowed between a point and the nearest line. Avoids snapping points to lines far away.}
}
\value{
\code{sf} object.
}
\description{
Function to "snap" spatial points to the nearest spatial line.
}
\author{
Shona Wilde
}
