% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_weighted_linear_regression.R
\name{st_weighted_linear_regression}
\alias{st_weighted_linear_regression}
\title{Function to perform distance-weighted quantile regression}
\usage{
st_weighted_linear_regression(st, location, formula, sigma = 250, verbose = T)
}
\arguments{
\item{st}{Object of class \code{sf} containing the data on which to perform the regression.}

\item{location}{Object of class \code{sf} containing POINT geometries of locations to perform the regression at.}

\item{formula}{A formula specifying the model, as in \code{\link[quantreg]{rq}} and \code{\link[stats]{lm}}.}

\item{sigma}{Width of the Gaussian kernel smoothing function.
Small \eqn{\sigma} leads to a narrow Gaussian and hence allows the user to focus on very localised effects, whereas
a bigger \eqn{\sigma} has the effect of smoothing things out.}

\item{verbose}{Should a message be displayed to the user?}
}
\value{
Tibble containing model results.
}
\description{
Function to perform distance-weighted quantile regression
}
\author{
Shona Wilde
}
