% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_weighted_mean.R
\name{st_weighted_mean}
\alias{st_weighted_mean}
\title{Distance weighted mean}
\usage{
st_weighted_mean(st, location, variable, sigma, verbose = T)
}
\arguments{
\item{st}{Object of class \code{sf} containing the data on which to perform the calculation}

\item{location}{Object of class \code{sf} containing POINT geometries of locations to perform the calculation at.}

\item{variable}{Name of the variable within \code{st} to use for the mean calculation.}

\item{sigma}{Width of the Gaussian kernel smoothing function in m.
Small \eqn{\sigma} leads to a narrow Gaussian and hence allows the user to focus on very localised effects, whereas
a bigger \eqn{\sigma} has the effect of smoothing things out.}

\item{verbose}{Should a message be displayed to the user?}
}
\value{
Tibble.
}
\description{
Function to calculate a geographically-weighted mean
}
\author{
Shona Wilde
}
