% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_weighted_quantile_regression.R
\name{st_weighted_quantile_regression}
\alias{st_weighted_quantile_regression}
\title{Distance-weighted regression}
\usage{
st_weighted_quantile_regression(
  st,
  location,
  formula,
  tau = c(0.5, 0.75, 0.9, 0.95, 0.99),
  sigma = 20,
  se = NULL,
  verbose = T
)
}
\arguments{
\item{st}{Object of class \code{sf} containing the data on which to perform the regression.}

\item{location}{Object of class \code{sf} containing POINT geometries of locations to perform the regression at.}

\item{formula}{A formula specifying the model, as in \code{\link[stats]{lm}} and \code{\link[quantreg]{rq}}.}

\item{tau}{The quantile(s) to be estimated. Generally a number between 0 and 1.}

\item{sigma}{Width of the Gaussian kernel smoothing function.
Small \eqn{\sigma} leads to a narrow Gaussian and hence allows the user to focus on very localised effects, whereas
a bigger \eqn{\sigma} has the effect of smoothing things out.}

\item{se}{Method to use for calculating standard errors. One of NULL, "nid", "rank", "iid", "ker", or "boot".}

\item{verbose}{Should a message be displayed to the user?}
}
\value{
Tibble containing model results.
}
\description{
Function to perform distance-weighted quantile regression
}
\author{
Shona Wilde
}
