% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_linear_regression.R
\name{weighted_linear_regression}
\alias{weighted_linear_regression}
\title{Function to perform weighted simple linear or multiple regression}
\usage{
weighted_linear_regression(data, formula, x = "x", mean_vector, sigma = 20)
}
\arguments{
\item{data}{Input data.}

\item{formula}{A formula specifying the model, as in \code{\link[stats]{lm}}.}

\item{x}{Variable for calculating weights.}

\item{mean_vector}{Vector of means for centre of Gaussian kernel. For example a vector of distances along a route.}

\item{sigma}{Width of the Gaussian kernel smoothing function.
Small \sigma leads to a narrow Gaussian and hence allows the user to focus on very localised effects, whereas
a bigger \sigma has the effect of smoothing things out.}
}
\value{
Tibble containing model results.
}
\description{
Function to perform weighted simple linear or multiple regression
}
\author{
Shona Wilde
}
