% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_quantile_regression.R
\name{weighted_quantile_regression}
\alias{weighted_quantile_regression}
\title{Function to perform weighted quantile regression}
\usage{
weighted_quantile_regression(
  data,
  formula,
  x = "x",
  mean_vector,
  tau = c(0.5, 0.75, 0.9, 0.95, 0.99),
  sigma = 20
)
}
\arguments{
\item{data}{Input data.}

\item{formula}{A formula specifying the model, as in \code{\link[quantreg]{rq}} and \code{\link[stats]{lm}}.}

\item{x}{Variable for calculating weights.}

\item{mean_vector}{Vector of means for centre of Gaussian kernel. For example a vector of distances along a route.}

\item{tau}{The quantile(s) to be estimated. Generally a number between 0 and 1.}

\item{sigma}{Width of the Gaussian kernel smoothing function.
Small \eqn{\sigma} leads to a narrow Gaussian and hence allows the user to focus on very localised effects, whereas
a bigger \eqn{\sigma} has the effect of smoothing things out.}
}
\value{
Tibble containing model results.
}
\description{
Function to perform weighted quantile regression
}
\author{
Shona Wilde
}
