% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hypotenuse_length.R
\name{calculate_hypotenuse_length}
\alias{calculate_hypotenuse_length}
\title{Function to calculate the length of the hypotenuse in a triangle}
\usage{
calculate_hypotenuse_length(x, y)
}
\arguments{
\item{x}{Horizontal side of triangle.}

\item{y}{Vertical side of triangle.}
}
\value{
Value.
}
\description{
Function to calculate the length of the hypotenuse in a triangle
}
\author{
Shona Wilde
}
