% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regression_line_length.R
\name{calculate_regression_line_length}
\alias{calculate_regression_line_length}
\title{Function to calculate the length of a regression line}
\usage{
calculate_regression_line_length(df, x, y)
}
\arguments{
\item{df}{Input data frame. Must contain a slope variable.}

\item{x}{x variable for regression.}

\item{y}{y variable for regression.}
}
\value{
Tibble.
}
\description{
Function to calculate the length of a regression line
}
\author{
Shona Wilde
}
