% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_plot_coloured.R
\name{leaflet_plot_coloured}
\alias{leaflet_plot_coloured}
\title{Function to plot coloured geometries on a leaflet map}
\usage{
leaflet_plot_coloured(
  st,
  colour_by = "value",
  palette = "plasma",
  popup = NULL,
  opacity = 0.5,
  fill_opacity = 0.2
)
}
\arguments{
\item{st}{Object with class \code{sf} or \code{sfc}.}

\item{colour_by}{Variable to colour geometries by.}

\item{palette}{Colours that values will be mapped to.}

\item{popup}{Vector of variables to be used as a pop-up on the map.}

\item{opacity}{Stroke opacity (edge of geometry).}

\item{fill_opacity}{Fill opacity (inside of geometry).}
}
\value{
Leaflet map.
}
\description{
This function is a wrapper around \code{\link[gissr]{leaflet_plot}} that allows geometries to be coloured according to a variable.
Discrete and continuous scales are supported.
}
\seealso{
\code{\link[gissr]{leaflet_plot}}, \code{\link[leaflet]{colorNumeric}}
}
\author{
Shona Wilde
}
